+++ indexer/sql/src/lib.rs
@@ -1,8 +1,9 @@
-pub mod active_model;
+mod active_model;
 pub mod block;
-pub mod context;
+mod context;
 pub mod entity;
-pub mod error;
+mod error;
+mod indexer_path;
 pub mod non_blocking_indexer;
 
-pub use context::Context;
+use context::Context;
